% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_laplace.R
\name{laplace_beta_names}
\alias{laplace_beta_names}
\title{Column names of the fixed effects of one process}
\usage{
laplace_beta_names(X, prefix = NULL)
}
\description{
Column names of the fixed effects of one process
}
\keyword{internal}
