% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tratio.R
\name{print_param_summary}
\alias{print_param_summary}
\title{Print categorized parameter summary}
\usage{
print_param_summary(summ, family_name, ci_names, fixed = character(0))
}
\description{
Print categorized parameter summary
}
\keyword{internal}
