% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/tulpaRatio-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_draws}
\alias{pp_check}
\alias{posterior_predict}
\alias{diagnostics}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayesplot}{\code{\link[bayesplot:pp_check]{pp_check()}}}

  \item{posterior}{\code{\link[posterior:as_draws]{as_draws()}}}

  \item{tulpa}{\code{\link[tulpa:diagnostics]{diagnostics()}}, \code{\link[tulpa:posterior_predict]{posterior_predict()}}}
}}

