% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rsr_basis}
\alias{rsr_basis}
\title{Restricted-covariate basis of a spatial specification}
\usage{
rsr_basis(spatial, data)
}
\arguments{
\item{spatial}{A spatial specification}

\item{data}{Data frame}
}
\value{
The basis from \code{\link[=rsr_basis_of]{rsr_basis_of()}} for a \code{spatial_rsr()}
specification, \code{NULL} otherwise
}
\description{
Restricted-covariate basis of a spatial specification
}
\keyword{internal}
